Add a function that outputs values of a file from a csv file.
import the csv file creator to create a csv file and write to it using a list of string values.

Use a try exepection to only read content after a file has been created sucessfully and then call the file reader.

Read the values and print them to the screen. Print any errrors if the file file fails to be created for some reason